chore: Move and extend API types
